Add TrackControl.js
