+++ src/TrackRowInfo.scss
@@ -7,53 +7,3 @@
     width: auto;
     position: relative;
 }
-
-.chgw-control {
-    display: "flex";
-    flex-direction: "row";
-    position: absolute;
-    border-radius: 2.5px;
-    background: white;
-    opacity: 0.6;
-    box-shadow: 0 0 0 1px rgba(0,0,0,.05),0 0 3px 0 rgba(0,0,0,.1);
-}
-
-.chgw-control:hover {
-    opacity: 1;
-}
-
-.chgw-button-sm {
-    height: 20px;
-    width: 20px;
-    padding: 4px;
-    cursor: pointer;
-}
-
-.chgw-button-sm:hover {
-    color: #fff;
-    background: #337ab7;
-    opacity: 1;
-}
-
-.chgw-button-left {
-    border-radius: 2.5px 0 0 2.5px;
-}
-.chgw-button-right {
-    border-radius: 0 2.5px 2.5px 0;
-}
-.chgw-button-top {
-    border-radius: 2.5px 2.5px 0 0;
-}
-.chgw-button-bottom {
-    border-radius: 0 0 2.5px 2.5px;
-}
-
-.chgw-button-disabled {
-    background: gray;
-    cursor: no-drop;
-}
-
-.chgw-button-disabled:hover {
-    color: black;
-    background: gray !important;
-}
